refactor(gui): modularise util and paste parsing
Split the monolithic gui/util.py into specialized submodules under the
gui/utils/ package. Refactored primer paste handling by extracting the
shared sequence parsing logic to a separate helper module, and updated
all imports to use the direct submodule paths.
